Updating sample file
