+++ lib/addons/index.ts
@@ -1,0 +1,7 @@
+export { AppMeshAddon } from './appmesh'
+export { ArgoCDAddon } from './argocd'
+export { CalicoAddon } from './calico'
+export { ContainerInsightsAddOn } from './cloudwatch'
+export { ClusterAutoScalerAddon } from './cluster-autoscaler'
+export { MetricsServerAddon } from './metrics-server'
+export { NginxAddon } from './nginx'
