+++ lib/addons/nginx/index.ts
@@ -2,11 +2,11 @@ import { ClusterAddOn, ClusterInfo } from "../../eksBlueprintStack";
 
 export class NginxAddon implements ClusterAddOn {
 
-  deploy(clusterInfo: ClusterInfo): void {
-       clusterInfo.cluster.addHelmChart("ngninx-ingress", {
+    deploy(clusterInfo: ClusterInfo): void {
+        clusterInfo.cluster.addHelmChart("ngninx-addon", {
             chart: "nginx-ingress",
             repository: "https://helm.nginx.com/stable",
             namespace: "kube-system"
         });
-  }
+    }
 }
