+++ lib/index.ts
@@ -1,8 +1,3 @@
 export * from './eksBlueprintStack';
-export * from './addons/appmesh/appMeshAddon';
-export * from './addons/argocd/argoCDAddon';
-export * from './addons/calico/calicoAddon';
-export * from './addons/cloudwatch/containerInsightsAddon';
-export * from './addons/cluster-autoscaler/clusterAutoscalerAddon';
-export * from './addons/metrics-server/metricsServerAddon';
-export * from './addons/nginx/nginxAddon';
+export * as Addons from './addons';
+
